implemented createRandomField
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@21911 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 64935d298b2713c13690e8996f900b93e86b6816

+++ sandbox/object_segmentation/include/object_segmentation/image_stereo/table_object_rf.h
@@ -85,9 +85,9 @@ class TableObjectRF
     void
     downsampleStereoCloud(sensor_msgs::PointCloud& full_stereo_cloud,
                           sensor_msgs::PointCloud& ds_stereo_cloud,
-                          const unsigned int voxel_x,
-                          const unsigned int voxel_y,
-                          const unsigned int voxel_z,
+                          const double voxel_x,
+                          const double voxel_y,
+                          const double voxel_z,
                           vector<unsigned int>& ds_labels,
                           map<unsigned int, pair<unsigned int, unsigned int> >& ds_idx2img_coords);
